requirements.txt
Django==3.2.9
djangorestframework==3.14.0
markdown==3.5.2
django-filter==23.5
pymongo==4.6.1
Werkzeug
djangorestframework-simplejwt